+++ src/components/layouts/PageHeader.tsx
@@ -1,21 +1,10 @@
-import { FC, PropsWithChildren, useState } from 'react';
-import { Search, PersonCircle, BoxArrowInRight, Cart } from 'react-bootstrap-icons';
+import { FC, PropsWithChildren } from 'react';
+import { PersonCircle, BoxArrowInRight, Cart } from 'react-bootstrap-icons';
 import { Link } from 'react-router-dom';
-import { Box, Image, Flex, Spacer, HStack, Text, Center, Input } from '@chakra-ui/react';
+import { Box, Image, Flex, Spacer, HStack, Text, Center } from '@chakra-ui/react';
+import SearchTerm from '@/components/SearchTerm';
 
 const PageHeader: FC<PropsWithChildren> = () => {
-  const [isSearchBarVisible, setIsSearchBarVisible] = useState(false);
-  const [searchQuery, setSearchQuery] = useState('');
-
-  const toggleSearchBar = () => {
-    setIsSearchBarVisible(!isSearchBarVisible);
-  };
-
-  const handleSearch = () => {
-    // TODO: 검색 기능을 수행하고, 다른 페이지로 리다이렉션
-    console.log('검색어:', searchQuery);
-  };
-
   return (
     <Box
       as="header"
@@ -71,11 +60,8 @@ const PageHeader: FC<PropsWithChildren> = () => {
         <Spacer />
 
         <HStack spacing={10} mr={8} align="center" pr={20}>
-          <Text fontWeight="bold" letterSpacing="0.1em" onClick={toggleSearchBar} cursor="pointer">
-            <Center>
-              <Search /> SEARCH
-            </Center>
-          </Text>
+          <SearchTerm />
+
           <Link to="/login">
             <Text fontWeight="bold" letterSpacing="0.1em" _hover={{ textDecoration: 'underline' }}>
               <Center>
@@ -99,34 +85,6 @@ const PageHeader: FC<PropsWithChildren> = () => {
           </Link>
         </HStack>
       </Flex>
-      {isSearchBarVisible && (
-        <Box
-          p={4}
-          bg="white"
-          position={{ base: 'absolute', md: 'fixed' }}
-          top={{ base: 0, md: '100px' }}
-          left={0}
-          right={0}
-          boxShadow="sm"
-          zIndex={1}
-          w="50%"
-          mx="auto"
-        >
-          <Input
-            placeholder="검색어를 입력해주세요."
-            size="sm"
-            borderColor="transparent"
-            borderBottomWidth="2px"
-            borderBottomColor="gray.600"
-            _hover={{ borderColor: 'gray.600' }}
-            value={searchQuery}
-            onChange={e => setSearchQuery(e.target.value)}
-            onKeyDown={e => e.key === 'Enter' && handleSearch()}
-            minWidth="500px"
-            w="100%"
-          />
-        </Box>
-      )}
     </Box>
   );
 };
