+++ src/pages/search/page.tsx
@@ -1,0 +1,13 @@
+import { Box } from '@chakra-ui/react';
+
+const SearchPage = () => {
+  return (
+    <Box p="10px 10%" bg="purple.50">
+      <Box bgColor="white" fontSize="xl" minH="1000px">
+        dd
+      </Box>
+    </Box>
+  );
+};
+
+export default SearchPage;
